pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = ""
authors = ["Anton Sidarevich <as@apostro.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
apache-airflow = "^2.6.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
